Moves CreateMetrics into the program-runtime crate.
