+++ program-runtime/src/lib.rs
@@ -11,6 +11,7 @@ extern crate solana_metrics;
 
 pub mod accounts_data_meter;
 pub mod compute_budget;
+pub mod executor;
 pub mod executor_cache;
 pub mod invoke_context;
 pub mod log_collector;
